projects/ai_assistant_with_langchain_on_local_machine/requirements.txt
python-dotenv
langchain_ollama
streamlit
